separating edx and caliper logs
