--- openedx_caliper_tracking/pavelib/__init__.py
@@ -1,1 +1,0 @@
-from . import tests
