--- openedx_caliper_tracking/pavelib/settings.py
@@ -1,6 +1,0 @@
-"""Settings related to or required by all paver tasks"""
-import os
-
-
-CALIPER_TESTS_API_ADDRESS = os.environ.get('CALIPER_TESTS_API')
-CALIPER_TESTS_API_TOKEN = os.environ.get('CALIPER_TESTS_API_TOKEN')
